+++ src/FhirStore.R4/Subscriptions/SubscriptionConverter.cs
@@ -53,6 +53,23 @@ public class SubscriptionConverter
         _maxSubscriptionTicks = TimeSpan.FromMinutes(MaxSubscriptionExpirationMinutes).Ticks;
     }
 
+    public void UpdateResourceStatus(object subscription, string statusLiteral)
+    {
+        if ((subscription == null) ||
+            (subscription is not Hl7.Fhir.Model.Subscription sub))
+        {
+            return;
+        }
+
+        Subscription.SubscriptionStatus? status = Hl7.Fhir.Utility.EnumUtility.ParseLiteral<Subscription.SubscriptionStatus>(statusLiteral);
+        if (status == null)
+        {
+            return;
+        }
+
+        sub.Status = status.Value;
+    }
+
     /// <summary>Attempts to parse a ParsedSubscription from the given object.</summary>
     /// <param name="subscription">The subscription.</param>
     /// <param name="common">      [out] The common.</param>
@@ -76,10 +93,15 @@ public class SubscriptionConverter
             expirationTicks = _maxSubscriptionTicks;
         }
 
+        HashSet<string> tags = (sub.Meta?.Tag?.Any() ?? false)
+            ? tags = new(sub.Meta.Tag.Select(t => $"{t.System}|{t.Code}"))
+            : [];
+
         common = new()
         {
             Id = sub.Id,
             TopicUrl = sub.Criteria,
+            Tags = tags,
             ChannelSystem = string.Empty,
             ChannelCode = sub.Channel.Type == null
                 ? string.Empty
@@ -222,7 +244,7 @@ public class SubscriptionConverter
     }
 
 
-    /// <summary>Attempts to parse a ParsedSubscription from the given object.</summary>
+    /// <summary>Attempts to create a Subscription from a ParsedSubscription object.</summary>
     /// <returns>True if it succeeds, false if it fails.</returns>
     public bool TryParse(ParsedSubscription common, out Subscription subscription)
     {
@@ -296,8 +318,23 @@ public class SubscriptionConverter
 
         subscription.Channel.PayloadElement.AddExtension(_urlBackport + _content, new Code(common.ContentLevel));
 
+        // add tags
+        if (common.Tags.Count != 0)
+        {
+            if (subscription.Meta == null)
+            {
+                subscription.Meta = new();
+            }
+
+            foreach (string tag in common.Tags)
+            {
+                string[] parts = tag.Split('|');
+                subscription.Meta.Tag.Add(new Coding(parts[0], parts[1]));
+            }
+        }
+
         // add parameters
-        if (common.Parameters.Any())
+        if (common.Parameters.Count != 0)
         {
             List<string> headers = new();
 
@@ -310,7 +347,7 @@ public class SubscriptionConverter
         }
 
         // add filters
-        if (common.Filters.Any())
+        if (common.Filters.Count != 0)
         {
             foreach (List<ParsedSubscription.SubscriptionFilter> filters in common.Filters.Values)
             {