+++ src/FhirStore.R4B/Subscriptions/SubscriptionConverter.cs
@@ -52,6 +52,23 @@ public class SubscriptionConverter
         _maxSubscriptionTicks = TimeSpan.FromMinutes(MaxSubscriptionExpirationMinutes).Ticks;
     }
 
+    public void UpdateResourceStatus(object subscription, string statusLiteral)
+    {
+        if ((subscription == null) ||
+            (subscription is not Hl7.Fhir.Model.Subscription sub))
+        {
+            return;
+        }
+
+        SubscriptionStatusCodes? status = Hl7.Fhir.Utility.EnumUtility.ParseLiteral<SubscriptionStatusCodes>(statusLiteral);
+        if (status == null)
+        {
+            return;
+        }
+
+        sub.Status = status.Value;
+    }
+
     /// <summary>Attempts to parse a ParsedSubscription from the given object.</summary>
     /// <param name="subscription">The subscription.</param>
     /// <param name="common">      [out] The common.</param>
@@ -75,10 +92,15 @@ public class SubscriptionConverter
             expirationTicks = _maxSubscriptionTicks;
         }
 
+        HashSet<string> tags = (sub.Meta?.Tag?.Any() ?? false)
+            ? tags = new(sub.Meta.Tag.Select(t => $"{t.System}|{t.Code}"))
+            : [];
+
         common = new()
         {
             Id = sub.Id,
             TopicUrl = sub.Criteria,
+            Tags = tags,
             ChannelSystem = string.Empty,
             ChannelCode = sub.Channel.Type == null
                 ? string.Empty
@@ -223,7 +245,7 @@ public class SubscriptionConverter
         return true;
     }
 
-    /// <summary>Attempts to parse a ParsedSubscription from the given object.</summary>
+    /// <summary>Attempts to create a Subscription from a ParsedSubscription object.</summary>
     /// <returns>True if it succeeds, false if it fails.</returns>
     public bool TryParse(ParsedSubscription common, out Subscription subscription)
     {
@@ -297,8 +319,23 @@ public class SubscriptionConverter
 
         subscription.Channel.PayloadElement.AddExtension(_contentUrl, new Code(common.ContentLevel));
 
+        // add tags
+        if (common.Tags.Count != 0)
+        {
+            if (subscription.Meta == null)
+            {
+                subscription.Meta = new();
+            }
+
+            foreach (string tag in common.Tags)
+            {
+                string[] parts = tag.Split('|');
+                subscription.Meta.Tag.Add(new Coding(parts[0], parts[1]));
+            }
+        }
+
         // add parameters
-        if (common.Parameters.Any())
+        if (common.Parameters.Count != 0)
         {
             List<string> headers = new();
 
@@ -311,7 +348,7 @@ public class SubscriptionConverter
         }
 
         // add filters
-        if (common.Filters.Any())
+        if (common.Filters.Count != 0)
         {
             foreach (List<ParsedSubscription.SubscriptionFilter> filters in common.Filters.Values)
             {